requirements.txt
aniso8601==7.0.0
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.4.0
backports.zoneinfo==0.2.1
colorama==0.4.4
Django==3.2.11
django-filter==21.1
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
Faker==11.3.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
iniconfig==1.1.1
mixer==7.2.0
packaging==21.3
pluggy==1.0.0
promise==2.3
psycopg2-binary==2.9.3
py==1.11.0
PyJWT==1.7.1
pyparsing==3.0.6
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tzdata==2021.5
